requirements.txt
asgiref==3.5.0
Babel==2.9.1
crispy-bootstrap5==0.6
Django==4.0.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-phonenumber-field==6.1.0
form==0.0.1
phonenumbers==8.12.45
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
